[BE] refactor :recycle: Issue 구조 개선
- getIssue(), delete() 시 본인 것이 아니면 불가하도록 개선

#19

+++ BE/src/main/java/com/codesquad/issuetracker/issue/business/IssueService.java
@@ -21,7 +21,13 @@ public class IssueService {
   }
 
   public Issue getIssue(User user, Long issueId) {
-    return issueRepository.findById(issueId).orElseThrow(NoSuchElementException::new);
+    Issue findIssue = issueRepository.findById(issueId).orElseThrow(NoSuchElementException::new);
+
+    if (!findIssue.isSameUser(user)) {
+      throw new RuntimeException("다른 유저의 Issue 입니다");
+    }
+
+    return findIssue;
   }
 
   @Transactional
@@ -32,6 +38,11 @@ public class IssueService {
   @Transactional
   public void delete(User user, Long issueId) {
     Issue findIssue = issueRepository.findById(issueId).orElseThrow(NoSuchElementException::new);
+
+    if (!findIssue.isSameUser(user)) {
+      throw new RuntimeException("다른 유저의 Issue 입니다");
+    }
+
     issueRepository.delete(findIssue);
   }
 }
